+++ client/src/draw_context.rs
@@ -1,3 +1,4 @@
+use glium::Program;
 use glium::backend::Facade;
 use glium::texture::MipmapsOption::NoMipmap;
 use glium::texture::Texture2d;
@@ -8,14 +9,29 @@ use error::DrawContextCreationError;
 pub struct DrawContext {
     pub camera: Matrix,
     pub texture_for_object_picking: Texture2d,
+    pub fill_id_program: Program,
 }
 
 impl DrawContext {
     pub fn new<F>(display: &F, width: u32, height: u32) -> Result<DrawContext, DrawContextCreationError> where F: Facade {
         let texture = try!(Texture2d::empty_with_format(display, U8U8U8U8, NoMipmap, width, height));
+        let fill_id_program = try!(Program::from_source(display, r#"
+            #version 410
+            uniform mat4 matrix;
+            in vec3 position;
+            void main() {
+                gl_Position = matrix * vec4(position, 1.0);
+            }"#, r#"
+            #version 410
+            uniform vec4 id;
+            out vec4 color;
+            void main() {
+                color = id;
+            }"#, None));
         Ok(DrawContext {
             camera: Matrix::orthographic(width as f32/10.0, height as f32/10.0, 0.0, 1.0),
             texture_for_object_picking: texture,
+            fill_id_program: fill_id_program,
         })
     }
